+++ vscode_extension/src/constants.ts
@@ -32,6 +32,10 @@ export namespace PROJECTS {
     export const REMOVE_PROJECT = `${PROJECTS.CONFIG.KEY}.removeProject`;
     export const COMPILE = `${PROJECTS.CONFIG.KEY}.compile`;
     export const RECREATE = `${PROJECTS.CONFIG.KEY}.recreate`;
+    export const COMPILE_FOR_DEBUGGING = `${PROJECTS.CONFIG.KEY}.compileForDebugging`;
+    export const COMPILE_SELECTED_PROJECT_FOR_DEBUGGING = `${PROJECTS.CONFIG.KEY}.compileSelectedProjectForDebugging`;
+    export const COMPILE_ALL_IN_WORKSPACE_FOR_DEBUGGING = `${PROJECTS.CONFIG.KEY}.compileAllInWorkspaceForDebugging`;
+    export const COMPILE_ALL_IN_GROUP_PROJECT_FOR_DEBUGGING = `${PROJECTS.CONFIG.KEY}.compileAllInGroupProjectForDebugging`;
     export const REFRESH = `${PROJECTS.CONFIG.KEY}.refresh`;
     export const DISCOVER_PROJECT_PATHS = `${PROJECTS.CONFIG.KEY}.discoverProjectPaths`;
     export const SET_MANUAL_PATH = `${PROJECTS.CONFIG.KEY}.setManualPath`;